Extract CLI parser and use visibility-state flags
Move argparse builders and command handlers into parser.py so
main.py stays a thin entrypoint. Replace repo create --visibility
with mutually exclusive --public/--private/--internal flags.
