+++ tests/test_cli_smoke.py
@@ -1,5 +1,6 @@
 import pytest
-from github_rest_cli.main import build_parser, cli, __version__
+from github_rest_cli.main import cli, __version__
+from github_rest_cli.parser import build_parser
 
 
 def test_version_flag(capsys):
@@ -56,3 +57,28 @@ def test_repo_get_subcommand_help(capsys):
 
     assert exc_info.value.code == 0
     assert "--name" in capsys.readouterr().out
+
+
+def test_repo_create_defaults_to_public():
+    parser = build_parser()
+    args = parser.parse_args(["repo", "create", "--name", "my-repo"])
+
+    assert args.visibility == "public"
+
+
+def test_repo_create_private_flag():
+    parser = build_parser()
+    args = parser.parse_args(["repo", "create", "--name", "my-repo", "--private"])
+
+    assert args.visibility == "private"
+
+
+def test_repo_create_public_and_private_conflict(capsys):
+    parser = build_parser()
+
+    with pytest.raises(SystemExit) as exc_info:
+        parser.parse_args(
+            ["repo", "create", "--name", "my-repo", "--public", "--private"]
+        )
+
+    assert exc_info.value.code == 2
